+++ src/templates/spinner.js
@@ -1,3 +1,9 @@
 import { html } from "lit";
 
-export default (o={}) => html`<span class="spinner fa fa-spinner centered ${o.classes || ''}"/>`
+export default (o={}) => {
+    if (o.classes?.includes('hor_centered')) {
+        return html`<div class="spinner__container"><span class="spinner fa fa-spinner centered ${o.classes || ''}"/></div>`
+    } else {
+        return html`<span class="spinner fa fa-spinner centered ${o.classes || ''}"/>`
+    }
+}